:art: Rename Http namespace to Ui
fix #63
