+++ config/packages/doctrine.yaml
@@ -23,6 +23,7 @@ doctrine:
 when@test:
     doctrine:
         dbal:
+            logging: false
             # "TEST_TOKEN" is typically set by ParaTest
             dbname_suffix: '_test%env(default::TEST_TOKEN)%'
