--- src/Http/Form/ContactType.php
@@ -1,50 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Http\Form;
-
-use App\Domain\Contact\Model\Contact;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\EmailType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-final class ContactType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('name', TextType::class, [
-                'label' => false,
-                'attr' => [
-                    'placeholder' => 'placeholder.contact.name',
-                ],
-            ])
-            ->add('email', EmailType::class, [
-                'label' => false,
-                'attr' => [
-                    'placeholder' => 'placeholder.contact.email',
-                ],
-            ])
-            ->add('content', TextareaType::class, [
-                'label' => false,
-                'attr' => [
-                    'placeholder' => 'placeholder.contact.content',
-                ],
-            ])
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'label' => false,
-            'data_class' => Contact::class,
-            'empty_data' => new Contact(),
-            'translation_domain' => 'form',
-        ]);
-    }
-}
